Move `ComponentColumnDescriptor` and `TimeColumnDescriptor`
