+++ crates/store/re_sorbet/Cargo.toml
@@ -21,8 +21,9 @@ all-features = true
 
 [dependencies]
 re_format_arrow.workspace = true
-re_log.workspace = true
 re_log_types.workspace = true
+re_log.workspace = true
+re_types_core.workspace = true
 
 arrow.workspace = true
 thiserror.workspace = true
